+++ spec/factories/exports.rb
@@ -3,8 +3,8 @@
 FactoryBot.define do
   factory :export do
     name { 'export' }
-    url { 'exports/export.json' }
-    status { 1 }
+    status { :created }
+    format { :json }
     user
   end
 end